Extract fixture task tree into ll_fixtures, move test crates under test/
test/ll_fixtures — reusable library with run_pipeline() that exercises
all ll features (nesting, progress, data, tags, sync/async, detached).

test/ll_test_cli — thin binary that sets up ll_stdio and runs fixtures.
Now just 15 lines instead of 365.

Publishable crates (ll, ll_stdio, ll_wasm, ll_macros) stay at root.
Test crates live under test/.
