+++ test/ll_test_cli/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ll_test_cli"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+ll = { path = "../../ll", features = ["tokio"] }
+ll_stdio = { path = "../../ll_stdio" }
+ll_fixtures = { path = "../ll_fixtures" }
+anyhow = "1"
+tokio = { version = "1", features = ["full"] }
